fixed issue of Normal user shouldn't able to see accounts, contacts, leads, opportunities, cases assigned to other users, displaying all the records based on loggedin user profile, displaying no records message when there are no items, no displaying closed accounts & leads in create/edit pages (#210)
* added django storages, loading local or aws storages based on env variable, removed reset password option for loggedin user, fixed issue of account edit in account list, removed address text in leads create page, properly aligining attachments, fixed issue of unable to see assigned contacts in edit account page

* added source column in leads list page, fixed issue converting leads to accounts in leads view page, displaying billing address details in accounts view page, added more colors for tags

* fixed issue of Normal user shouldn't get edit & delete options under actions in document list page, displaying contacts in account detail page, APISettings modal updated with website field and also CRUD opearations views respective to website field

* fixed issue of Normal user shouldn't able to see accounts, contacts, leads, opportunities, cases assigned to other users, displaying all the records based on loggedin user profile, displaying no records message when there are no items, no displaying closed accounts & leads in create/edit pages

+++ contacts/views.py
@@ -14,6 +14,8 @@ from common.utils import COUNTRIES
 from contacts.models import Contact
 from contacts.forms import ContactForm, ContactCommentForm, ContactAttachmentForm
 from accounts.models import Account
+from django.core.exceptions import PermissionDenied
+from django.db.models import Q
 
 
 class ContactsListView(LoginRequiredMixin, TemplateView):
@@ -23,27 +25,42 @@ class ContactsListView(LoginRequiredMixin, TemplateView):
 
     def get_queryset(self):
         queryset = self.model.objects.all()
+        if self.request.user.role != "ADMIN" and not self.request.user.is_superuser:
+            queryset = queryset.filter(
+                Q(assigned_to__in=[self.request.user]) | Q(created_by=self.request.user))
+
         request_post = self.request.POST
         if request_post:
             if request_post.get('first_name'):
-                queryset = queryset.filter(first_name__icontains=request_post.get('first_name'))
+                queryset = queryset.filter(
+                    first_name__icontains=request_post.get('first_name'))
             if request_post.get('city'):
-                queryset = queryset.filter(address__city__icontains=request_post.get('city'))
+                queryset = queryset.filter(
+                    address__city__icontains=request_post.get('city'))
             if request_post.get('phone'):
-                queryset = queryset.filter(phone__icontains=request_post.get('phone'))
+                queryset = queryset.filter(
+                    phone__icontains=request_post.get('phone'))
             if request_post.get('email'):
-                queryset = queryset.filter(email__icontains=request_post.get('email'))
+                queryset = queryset.filter(
+                    email__icontains=request_post.get('email'))
+            if request_post.getlist('assigned_to'):
+                queryset = queryset.filter(
+                    assigned_to__id__in=request_post.getlist('assigned_to'))
         return queryset
 
     def get_context_data(self, **kwargs):
         context = super(ContactsListView, self).get_context_data(**kwargs)
         context["contact_obj_list"] = self.get_queryset()
         context["per_page"] = self.request.POST.get('per_page')
-
+        context["users"] = User.objects.filter(
+            is_active=True).order_by('email')
+        context["assignedto_list"] = [
+            int(i) for i in self.request.POST.getlist('assigned_to', []) if i]
         search = False
         if (
             self.request.POST.get('first_name') or self.request.POST.get('city') or
-            self.request.POST.get('phone') or self.request.POST.get('email')
+            self.request.POST.get('phone') or self.request.POST.get(
+                'email') or self.request.POST.get('assigned_to')
         ):
             search = True
         context["search"] = search
@@ -79,10 +96,9 @@ class CreateContactView(LoginRequiredMixin, CreateView):
             contact_obj.created_by = self.request.user
             contact_obj.save()
             if self.request.GET.get('view_account', None):
-                if Account.objects.filter(id=
-                    int(self.request.GET.get('view_account'))).exists():
-                    Account.objects.get(id=
-                        int(self.request.GET.get('view_account'))).contacts.add(contact_obj)
+                if Account.objects.filter(id=int(self.request.GET.get('view_account'))).exists():
+                    Account.objects.get(id=int(self.request.GET.get(
+                        'view_account'))).contacts.add(contact_obj)
             return self.form_valid(form)
 
         return self.form_invalid(form)
@@ -90,7 +106,8 @@ class CreateContactView(LoginRequiredMixin, CreateView):
     def form_valid(self, form):
         contact_obj = form.save(commit=False)
         if self.request.POST.getlist('assigned_to', []):
-            contact_obj.assigned_to.add(*self.request.POST.getlist('assigned_to'))
+            contact_obj.assigned_to.add(
+                *self.request.POST.getlist('assigned_to'))
             assigned_to_list = self.request.POST.getlist('assigned_to')
             current_site = get_current_site(self.request)
             for assigned_to_user in assigned_to_list:
@@ -109,9 +126,11 @@ class CreateContactView(LoginRequiredMixin, CreateView):
         if self.request.FILES.get('contact_attachment'):
             attachment = Attachments()
             attachment.created_by = self.request.user
-            attachment.file_name = self.request.FILES.get('contact_attachment').name
+            attachment.file_name = self.request.FILES.get(
+                'contact_attachment').name
             attachment.contact = contact_obj
-            attachment.attachment = self.request.FILES.get('contact_attachment')
+            attachment.attachment = self.request.FILES.get(
+                'contact_attachment')
             attachment.save()
 
         if self.request.is_ajax():
@@ -157,16 +176,21 @@ class ContactDetailView(LoginRequiredMixin, DetailView):
 
     def get_context_data(self, **kwargs):
         context = super(ContactDetailView, self).get_context_data(**kwargs)
-
+        user_assgn_list = [i.id for i in context['object'].assigned_to.all()]
+        if self.request.user == context['object'].created_by:
+            user_assgn_list.append(self.request.user.id)
+        if self.request.user.role != "ADMIN" and not self.request.user.is_superuser:
+            if self.request.user.id not in user_assgn_list:
+                raise PermissionDenied
         assigned_data = []
         for each in context['contact_record'].assigned_to.all():
             assigned_dict = {}
             assigned_dict['id'] = each.id
-            assigned_dict['name'] =  each.email
+            assigned_dict['name'] = each.email
             assigned_data.append(assigned_dict)
 
         context.update({"comments": context["contact_record"].contact_comments.all(),
-                        'attachments':context["contact_record"].contact_attachment.all(),
+                        'attachments': context["contact_record"].contact_attachment.all(),
                         "assigned_data": json.dumps(assigned_data)
                         })
         return context
@@ -206,8 +230,10 @@ class UpdateContactView(LoginRequiredMixin, UpdateView):
         all_members_list = []
         if self.request.POST.getlist('assigned_to', []):
             current_site = get_current_site(self.request)
-            assigned_form_users = form.cleaned_data.get('assigned_to').values_list('id', flat=True)
-            all_members_list = list(set(list(assigned_form_users)) - set(list(assigned_to_ids)))
+            assigned_form_users = form.cleaned_data.get(
+                'assigned_to').values_list('id', flat=True)
+            all_members_list = list(
+                set(list(assigned_form_users)) - set(list(assigned_to_ids)))
             if len(all_members_list):
                 for assigned_to_user in all_members_list:
                     user = get_object_or_404(User, pk=assigned_to_user)
@@ -218,22 +244,25 @@ class UpdateContactView(LoginRequiredMixin, UpdateView):
                         'protocol': self.request.scheme,
                         'contact': contact_obj
                     })
-                    email = EmailMessage(mail_subject, message, to=[user.email])
+                    email = EmailMessage(
+                        mail_subject, message, to=[user.email])
                     email.content_subtype = "html"
                     email.send()
 
             contact_obj.assigned_to.clear()
-            contact_obj.assigned_to.add(*self.request.POST.getlist('assigned_to'))
+            contact_obj.assigned_to.add(
+                *self.request.POST.getlist('assigned_to'))
         else:
             contact_obj.assigned_to.clear()
-
 
         if self.request.FILES.get('contact_attachment'):
             attachment = Attachments()
             attachment.created_by = self.request.user
-            attachment.file_name = self.request.FILES.get('contact_attachment').name
+            attachment.file_name = self.request.FILES.get(
+                'contact_attachment').name
             attachment.contact = contact_obj
-            attachment.attachment = self.request.FILES.get('contact_attachment')
+            attachment.attachment = self.request.FILES.get(
+                'contact_attachment')
             attachment.save()
         if self.request.POST.get('from_account'):
             from_account = self.request.POST.get('from_account')
@@ -244,7 +273,8 @@ class UpdateContactView(LoginRequiredMixin, UpdateView):
 
     def form_invalid(self, form):
         address_obj = self.object.address
-        address_form = BillingAddressForm(self.request.POST, instance=address_obj)
+        address_form = BillingAddressForm(
+            self.request.POST, instance=address_obj)
         if self.request.is_ajax():
             return JsonResponse({'error': True, 'contact_errors': form.errors,
                                  'address_errors': address_form.errors})
@@ -254,6 +284,13 @@ class UpdateContactView(LoginRequiredMixin, UpdateView):
     def get_context_data(self, **kwargs):
         context = super(UpdateContactView, self).get_context_data(**kwargs)
         context["contact_obj"] = self.object
+        user_assgn_list = [
+            i.id for i in context["contact_obj"].assigned_to.all()]
+        if self.request.user == context['contact_obj'].created_by:
+            user_assgn_list.append(self.request.user.id)
+        if self.request.user.role != "ADMIN" and not self.request.user.is_superuser:
+            if self.request.user.id not in user_assgn_list:
+                raise PermissionDenied
         context["address_obj"] = self.object.address
         context["contact_form"] = context["form"]
         context["users"] = self.users
@@ -280,13 +317,15 @@ class RemoveContactView(LoginRequiredMixin, View):
     def post(self, request, *args, **kwargs):
         contact_id = kwargs.get("pk")
         self.object = get_object_or_404(Contact, id=contact_id)
-        if self.object.address_id:
-            self.object.address.delete()
-        self.object.delete()
-        if self.request.is_ajax():
-            return JsonResponse({'error': False})
-
-        return redirect("contacts:list")
+        if self.request.user.role != "ADMIN" and not self.request.user.is_superuser and self.request.user != self.object.created_by:
+            raise PermissionDenied
+        else:
+            if self.object.address_id:
+                self.object.address.delete()
+            self.object.delete()
+            if self.request.is_ajax():
+                return JsonResponse({'error': False})
+            return redirect("contacts:list")
 
 
 class AddCommentView(LoginRequiredMixin, CreateView):
@@ -296,7 +335,8 @@ class AddCommentView(LoginRequiredMixin, CreateView):
 
     def post(self, request, *args, **kwargs):
         self.object = None
-        self.contact = get_object_or_404(Contact, id=request.POST.get('contactid'))
+        self.contact = get_object_or_404(
+            Contact, id=request.POST.get('contactid'))
         if (
             request.user in self.contact.assigned_to.all() or
             request.user == self.contact.created_by or request.user.is_superuser or
@@ -305,7 +345,7 @@ class AddCommentView(LoginRequiredMixin, CreateView):
             form = self.get_form()
             if form.is_valid():
                 return self.form_valid(form)
-    
+
             return self.form_invalid(form)
 
         data = {'error': "You don't have permission to comment."}
@@ -330,14 +370,15 @@ class UpdateCommentView(LoginRequiredMixin, View):
     http_method_names = ["post"]
 
     def post(self, request, *args, **kwargs):
-        self.comment_obj = get_object_or_404(Comment, id=request.POST.get("commentid"))
+        self.comment_obj = get_object_or_404(
+            Comment, id=request.POST.get("commentid"))
         if request.user == self.comment_obj.commented_by:
             form = ContactCommentForm(request.POST, instance=self.comment_obj)
             if form.is_valid():
                 return self.form_valid(form)
-            
+
             return self.form_invalid(form)
-    
+
         data = {'error': "You don't have permission to edit this comment."}
         return JsonResponse(data)
 
@@ -356,12 +397,13 @@ class UpdateCommentView(LoginRequiredMixin, View):
 class DeleteCommentView(LoginRequiredMixin, View):
 
     def post(self, request, *args, **kwargs):
-        self.object = get_object_or_404(Comment, id=request.POST.get("comment_id"))
+        self.object = get_object_or_404(
+            Comment, id=request.POST.get("comment_id"))
         if request.user == self.object.commented_by:
             self.object.delete()
             data = {"cid": request.POST.get("comment_id")}
             return JsonResponse(data)
-        
+
         data = {'error': "You don't have permission to delete this comment."}
         return JsonResponse(data)
 
@@ -384,7 +426,8 @@ class AddAttachmentsView(LoginRequiredMixin, CreateView):
 
     def post(self, request, *args, **kwargs):
         self.object = None
-        self.contact = get_object_or_404(Contact, id=request.POST.get('contactid'))
+        self.contact = get_object_or_404(
+            Contact, id=request.POST.get('contactid'))
         if (
                 request.user in self.contact.assigned_to.all() or
                 request.user == self.contact.created_by or request.user.is_superuser or
@@ -393,9 +436,9 @@ class AddAttachmentsView(LoginRequiredMixin, CreateView):
             form = self.get_form()
             if form.is_valid():
                 return self.form_valid(form)
-            
+
             return self.form_invalid(form)
-        
+
         data = {'error': "You don't have permission to add attachment."}
         return JsonResponse(data)
 
@@ -410,7 +453,7 @@ class AddAttachmentsView(LoginRequiredMixin, CreateView):
             "attachment": attachment.file_name,
             "attachment_url": attachment.attachment.url,
             "created_on": attachment.created_on,
-            "download_url": reverse('common:download_attachment', kwargs={'pk':attachment.id}),
+            "download_url": reverse('common:download_attachment', kwargs={'pk': attachment.id}),
             "attachment_display": attachment.get_file_type_display(),
             "created_by": attachment.created_by.email
         })
@@ -422,7 +465,8 @@ class AddAttachmentsView(LoginRequiredMixin, CreateView):
 class DeleteAttachmentsView(LoginRequiredMixin, View):
 
     def post(self, request, *args, **kwargs):
-        self.object = get_object_or_404(Attachments, id=request.POST.get("attachment_id"))
+        self.object = get_object_or_404(
+            Attachments, id=request.POST.get("attachment_id"))
         if (
             request.user == self.object.created_by or request.user.is_superuser or
             request.user.role == 'ADMIN'
@@ -430,6 +474,7 @@ class DeleteAttachmentsView(LoginRequiredMixin, View):
             self.object.delete()
             data = {"aid": request.POST.get("attachment_id")}
             return JsonResponse(data)
-        
+
         data = {'error': "You don't have permission to delete this attachment."}
         return JsonResponse(data)
+
